fmpz_mod_poly set
